Address PR #49 review feedback from all reviewers
Bug fixes:
- Move density nudge after soft knee to prevent negative strength squaring
  to a false positive (0→-0.1→0.025 instead of 0). All reviewers flagged.
- Initialize onsetDensityWindowStart_ to time_.millis() in begin() to avoid
  huge first elapsed time. Copilot + Gemini flagged.
- Swap call order: updateOnsetDensity() before synthesizeRhythmStrength()
  so rhythmStrength uses fresh density value. Copilot + Claude flagged.

Naming:
- Rename perBandBassThreshMult_ to perBandThreshMult_ since it applies to
  both bass AND mid thresholds, not just bass. Claude flagged.

Minor:
- Change onsetCountInWindow_ from int to uint16_t (embedded convention)
- Remove stale cppcheck-suppress variableScope inline comment
- Fix IOI smoothing comment to accurately describe the in-place algorithm
- Update FT CPU estimate in IMPROVEMENT_PLAN.md (Goertzel O(lags×samples),
  not single FFT as originally planned)
- Update BandWeightedFluxDetector CPU estimate for history shift overhead

+++ blinky-things/audio/AudioController.h
@@ -538,7 +538,7 @@ private:
 
     // Onset density tracking
     float onsetDensity_ = 0.0f;            // Smoothed onsets/second (EMA)
-    int onsetCountInWindow_ = 0;           // Onsets detected in current 1s window
+    uint16_t onsetCountInWindow_ = 0;      // Onsets detected in current 1s window
     uint32_t onsetDensityWindowStart_ = 0; // Window start timestamp (ms)
 
     // IOI histogram onset ring buffer (records sample indices of detected onsets)
